logging/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pytest==8.3.3
httpx==0.27.2
ruff==0.7.0

fluent-logger==0.11.0
